Re-modularized the files
