+++ lib/canonicalization.ts
@@ -1,44 +1,21 @@
 import * as rdf from 'rdf-js';
-import { GlobalState, BNodeId, Hash, Graph, NDegreeHashResult, QuadToNquad, Logger, NopLogger } from './types';
-import { IdIssuer, compute_n_degree_hash, compute_first_degree_hash, get_bnodeid } from './utils';
+import { GlobalState, BNodeId, Hash, Graph, NDegreeHashResult, get_bnodeid } from './common';
+import { compute_first_degree_hash }                                         from './hash_1_degree_quads';
+import { compute_n_degree_hash }                                             from './hash_n_degree_quads';
+import { IdIssuer }                                                          from './issue_identifier';
 
-export class URDNA2015 {
-    private _state:    GlobalState;
 
-    /**
-     * 
-     * @param data_factory  An implementation of the generic RDF DataFactory interface, see http://rdf.js.org/data-model-spec/#datafactory-interface
-     * @param quad_to_nquad A function that converts an rdf.Quad into a bona fide nquad string
-     */
-    constructor(data_factory: rdf.DataFactory, quad_to_nquad: QuadToNquad, logger: Logger = new NopLogger() ) {
-        this._state = {
-            bnode_to_quads   : {},
-            // This will map a calculated hash value to the bnodes it characterizes. In
-            // simple cases it is a 1-1 mapping but, sometimes, it is a 1-many.
-            // This structure is filled in step 5
-            hash_to_bnodes   : {},
-            canonical_issuer : new IdIssuer(),
-            data_factory     : data_factory,
-            quad_to_nquad    : quad_to_nquad,
-            logger           : logger
-        }
-    }
-
-    private initialize() {
-        this._state.bnode_to_quads   = {};
-        this._state.hash_to_bnodes   = {};
-        this._state.canonical_issuer = new IdIssuer();
-    }
-
-    /**
-     * Implementation of the main algorithmic steps
-     * 
-     * @param input_dataset 
-     * @returns 
-     */
-    canonicalize(input_dataset: Graph): Graph {
+/**
+ * Implementation of the main algorithmic steps
+ * 
+ * @param input_dataset 
+ * @returns 
+ */
+export function compute_canonicalized_graph(state: GlobalState, input_dataset: Graph): Graph {
         // Re-initialize the state information: canonicalization should always start with a clean state
-        this.initialize();
+        state.bnode_to_quads   = {};
+        state.hash_to_bnodes   = {};
+        state.canonical_issuer = new IdIssuer();
 
         const retval: Graph = new Set();
 
@@ -50,10 +27,10 @@ export class URDNA2015 {
                 const bnode_map = (t: rdf.Term): void => {
                     if (t.termType === "BlankNode") {
                         const bnode: BNodeId = get_bnodeid(t);
-                        if (this._state.bnode_to_quads[bnode] === undefined) {
-                            this._state.bnode_to_quads[bnode] = [quad];
+                        if (state.bnode_to_quads[bnode] === undefined) {
+                            state.bnode_to_quads[bnode] = [quad];
                         } else {
-                            this._state.bnode_to_quads[bnode].push(quad);
+                            state.bnode_to_quads[bnode].push(quad);
                         }
                     }
                 }
@@ -63,24 +40,24 @@ export class URDNA2015 {
             }
         }
 
-        this._state.logger.debug(`§4.5.3 (2) bnode to quads: ${JSON.stringify(this._state.bnode_to_quads,null,4)}`);
+        state.logger.debug(`§4.5.3 (2) bnode to quads: ${JSON.stringify(state.bnode_to_quads,null,4)}`);
 
         // Step 3
         {
             // Compute a hash value for each bnode (depending on the quads it appear in)
             // In simple cases a hash value refers to one bnode only; in unlucky cases there
             // may be more. Hence the usage of the hash_to_bnodes map.
-            Object.keys(this._state.bnode_to_quads).forEach((n: BNodeId): void => {
+            Object.keys(state.bnode_to_quads).forEach((n: BNodeId): void => {
                 // Step 3.1
-                const hfn: Hash = compute_first_degree_hash(this._state, n)
+                const hfn: Hash = compute_first_degree_hash(state, n)
                 // Step 3.2
-                if (this._state.hash_to_bnodes[hfn] === undefined) {
-                    this._state.hash_to_bnodes[hfn] = [n];
+                if (state.hash_to_bnodes[hfn] === undefined) {
+                    state.hash_to_bnodes[hfn] = [n];
                 } else {
-                    this._state.hash_to_bnodes[hfn].push(n);
+                    state.hash_to_bnodes[hfn].push(n);
                 }
             });
-            this._state.logger.info(`§4.5.3 (3) hash to bnodes: \n${this._state.hash_to_bnodes}`)
+            state.logger.info(`§4.5.3 (3) hash to bnodes: \n${state.hash_to_bnodes}`)
         }
 
         // Step 4
@@ -91,9 +68,9 @@ export class URDNA2015 {
             // It is important to order the hashes, because it influences the order of issuing the canonical ids.
             // If a bnode is "handled", ie, it does have a canonical ID, it is removed from the
             // state structure on hash->bnodes. 
-            const hashes: Hash[] = Object.keys(this._state.hash_to_bnodes).sort();
+            const hashes: Hash[] = Object.keys(state.hash_to_bnodes).sort();
             for (const hash of hashes) {
-                const identifier_list: BNodeId[] = this._state.hash_to_bnodes[hash];
+                const identifier_list: BNodeId[] = state.hash_to_bnodes[hash];
                 // Step 4.1
                 // Filter out the nasty cases
                 if (identifier_list.length > 1) continue;
@@ -103,12 +80,12 @@ export class URDNA2015 {
                 // Note that the IdIssuer automatically stores the (existing, issued) pairs for the
                 // bnode identifier; these are retrieved in the last step when a new, normalized
                 // graph is created.
-                const canon_id = this._state.canonical_issuer.issue_id(identifier_list[0]);
-                this._state.logger.info(`§4.5.3 (4) canonicalization of ${identifier_list[0]} -> ${canon_id}`);
+                const canon_id = state.canonical_issuer.issue_id(identifier_list[0]);
+                state.logger.info(`§4.5.3 (4) canonicalization of ${identifier_list[0]} -> ${canon_id}`);
 
                 // Step 4.3
                 // Remove the corresponding hash
-                delete this._state.hash_to_bnodes[hash];
+                delete state.hash_to_bnodes[hash];
             }
         }
 
@@ -116,10 +93,10 @@ export class URDNA2015 {
         // This step takes care of the bnodes that do not have been canonicalized in the previous step,
         // because their simple, first degree hashes are not unique.
         {
-            const hashes: Hash[] = Object.keys(this._state.hash_to_bnodes).sort();
+            const hashes: Hash[] = Object.keys(state.hash_to_bnodes).sort();
             for (const hash of hashes) {
-                const identifier_list: BNodeId[] = this._state.hash_to_bnodes[hash];
-                this._state.logger.info(`§4.5.3 (5) identifier list with shared hashes: ${identifier_list} for hash: ${hash}`);
+                const identifier_list: BNodeId[] = state.hash_to_bnodes[hash];
+                state.logger.info(`§4.5.3 (5) identifier list with shared hashes: ${identifier_list} for hash: ${hash}`);
                 // This cycle takes care of all problematic cases that share the same hash
                 // Step 5.1
                 // This stores a calculated hash and its relates identifier issuer for each
@@ -128,7 +105,7 @@ export class URDNA2015 {
 
                 // Step 5.2
                 for (const n of identifier_list) {
-                    if (this._state.canonical_issuer.is_set(n)) {
+                    if (state.canonical_issuer.is_set(n)) {
                         // Step 5.2.1
                         continue;
                     } else {
@@ -137,13 +114,13 @@ export class URDNA2015 {
                         // Step 5.2.3
                         const bn = temporary_issuer.issue_id(n);
                         // Step 5.2.4
-                        const result: NDegreeHashResult = compute_n_degree_hash(this._state, n, temporary_issuer);
-                        this._state.logger.debug(`§4.5.3 (5.2.4) computed n-degree hash: ${JSON.stringify(result,null,4)}`);
+                        const result: NDegreeHashResult = compute_n_degree_hash(state, n, temporary_issuer);
+                        state.logger.debug(`§4.5.3 (5.2.4) computed n-degree hash: ${JSON.stringify(result,null,4)}`);
                         hash_path_list.push(result);
                     }
                 }
 
-                this._state.logger.info(`§4.5.3 (5.2) hash path list: ${JSON.stringify(hash_path_list,null,4)}`);
+                state.logger.info(`§4.5.3 (5.2) hash path list: ${JSON.stringify(hash_path_list,null,4)}`);
 
                 // Step 5.3
                 const ordered_hash_path_list = hash_path_list.sort((a,b): number => {
@@ -154,7 +131,7 @@ export class URDNA2015 {
                 for (const result of ordered_hash_path_list) {
                     // Step 5.3.1
                     for(const existing_identifier of result.issuer.existing_identifiers()) {
-                        this._state.canonical_issuer.issue_id(existing_identifier);
+                        state.canonical_issuer.issue_id(existing_identifier);
                     }
                 }
             }
@@ -165,9 +142,9 @@ export class URDNA2015 {
             // This function replaces the term with its canonical equivalent, if applicable
             const replace_bnode = (term: rdf.Term): rdf.Term => {
                 if (term.termType === "BlankNode") {
-                    const canonical = this._state.canonical_issuer.issue_id(get_bnodeid(term));
+                    const canonical = state.canonical_issuer.issue_id(get_bnodeid(term));
                     // Remove the `_:` before creating the new bnode...
-                    return this._state.data_factory.blankNode(canonical.slice(2))
+                    return state.data_factory.blankNode(canonical.slice(2))
                 } else {
                     return term;
                 }
@@ -177,12 +154,12 @@ export class URDNA2015 {
                 const subject_copy = replace_bnode(quad.subject) as rdf.Quad_Subject;
                 const object_copy  = replace_bnode(quad.object) as rdf.Quad_Object;
                 const graph_copy   = replace_bnode(quad.graph) as rdf.Quad_Graph;
-                retval.add(this._state.data_factory.quad(subject_copy,quad.predicate,object_copy,graph_copy))
+                retval.add(state.data_factory.quad(subject_copy,quad.predicate,object_copy,graph_copy))
             }
         }
 
         // Step 7
-        this._state.logger.debug(`§4.5.3 Leaving function\n${JSON.stringify(this._state,null,4)}`);
+        state.logger.debug(`§4.5.3 Leaving function\n${JSON.stringify(state,null,4)}`);
         return retval;
     }
-}
+
