+++ lib/common.ts
@@ -1,5 +1,6 @@
-import * as rdf     from 'rdf-js';
-import { IdIssuer } from './utils';
+import * as rdf       from 'rdf-js';
+import { createHash } from 'crypto';
+import { IdIssuer }   from './issue_identifier';
 
 export namespace Constants {
     /** The hashing algorithm's name used in the module */
@@ -42,7 +43,6 @@ export interface GlobalState extends C14nState {
     logger        : Logger;
 }
 
-
 export interface NDegreeHashResult {
     hash: Hash;
     issuer: IdIssuer
@@ -62,4 +62,38 @@ export class NopLogger implements Logger {
     info(message: string, ...otherData: any[]): void {};
 }
 
+
+/** Per spec, the BlankNode's "value" does not include the "_:", whereas it does in the C14 algorithm...  */
+export function get_bnodeid(term: rdf.BlankNode): BNodeId {
+    return `_:${term.value}`;
+}
+
+/**
+ * Return the hash of a string.
+ * 
+ * @param data 
+ * @returns hash value
+ */
+ export function compute_hash(data: string): Hash {
+    return createHash(Constants.HASH_ALGORITHM).update(data).digest('hex');
+}
+
+/**
+ * Return the hash of an array of nquad statements; per spec, this means
+ * concatenating all nquads into a long array. Care should be taken that each
+ * quad must end with a single `/n`.
+ * 
+ * @param nquads
+ * @returns hash value
+ * 
+ */
+export function hash_nquads(nquads: string[]): Hash {
+    // It may be better to add to a hash the quads one by one and then make the calculation;
+    // this is a possible optimization for another day...
+
+    // Care should be taken that the final data to be hashed include a single `/n`
+    // for every quad, before joining the quads into a string that must be hashed
+    const data: string = nquads.map((q:string): string => q.endsWith('\n') ? q : `${q}\n`).join('');
+    return compute_hash(data);
+}
