+++ tester/main.ts
@@ -1,7 +1,6 @@
 import * as rdf from 'rdf-js';
 import { Command }              from 'commander';
-import { URDNA2015 }            from '../lib/c14n';
-import { Graph }                from '../lib/types';
+import { URDNA2015, Graph }     from '../index';
 import { SimpleLogger, Levels } from './logger';
 import * as rdfn3               from './rdfn3';
 
@@ -29,11 +28,11 @@ async function main(): Promise<void> {
 
     logger.info(`Original graph: \n${rdfn3.graph_to_nquads(input).join('\n')}`);
 
-
-
-    const canonicalizer = new URDNA2015(rdfn3.DataFactory,rdfn3.quad_to_nquad,logger);
+    const canonicalizer = new URDNA2015(rdfn3.DataFactory, rdfn3.quad_to_nquad, logger);
     const normalized: Graph = canonicalizer.canonicalize(input);
-    logger.info(`Canonicalized graph: \n${rdfn3.graph_to_nquads(normalized).join('\n')}`);
+
+    const normalized_quads: string = rdfn3.graph_to_nquads(normalized).sort().join('\n');
+    console.log(`Canonicalized graph: \n${normalized_quads}`);
 }
 
 main()
